move add_location to preprocess
