+++ echopype/preprocess/__init__.py
@@ -1,3 +1,4 @@
 from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise
+from .data_coherence import add_location
 
-__all__ = ["compute_MVBS", "compute_MVBS_index_binning", "remove_noise"]
+__all__ = ["compute_MVBS", "compute_MVBS_index_binning", "remove_noise", "add_location"]
