+++ echopype/tests/utils/test_utils_common.py
@@ -102,22 +102,3 @@ def _check_swap(ds, ds_swap):
     assert "frequency_nominal" not in ds.dims
     assert "frequency_nominal" in ds_swap.dims
     assert "channel" not in ds_swap.dims
-
-
-def test_add_location(test_path):
-    ed = ep.open_raw(
-        test_path["EK60"] / "Winter2017-D20170115-T150122.raw",
-        sonar_model="EK60"
-    )
-    ds = ep.calibrate.compute_Sv(ed)
-
-    def _check_var(ds_test):
-        assert "latitude" in ds_test
-        assert "longitude" in ds_test
-        assert "time1" not in ds_test
-
-    ds_all = ep.utils.common.add_location(ds=ds, echodata=ed)
-    _check_var(ds_all)
-
-    ds_sel = ep.utils.common.add_location(ds=ds, echodata=ed, nmea_sentence="GGA")
-    _check_var(ds_sel)
